Refactor to smaller funcs
